brain_service/requirements.txt
grpcio>=1.60.0
grpcio-tools>=1.60.0
pydantic>=2.0.0
httpx>=0.27.0
